Implement Slivka processing demo
- Implement prototype `process_slivka` method in FastaProcessor
- Utilize `process_slivka` in SubmissionHandler, replacing the previous `process` method
- Rename existing `process` method in FastaProcessor to `process_local`
- Add slivka-client to requirements.txt
- Add SLIVKA_URL configuration to config.py

The new `FastaProcessor.process_slivka` method acts as a synchronous Slivka submission handler that performs the following:
1) Instantiates a SlivkaClient
2) Builds and submits the submission payload
3) Waits for the job to complete
4) Downloads the results
5) Makes the selected result available for download

This change significantly extends the capabilities of this example application and lays the groundwork for fulfilling the design requirement of building an extensible template application for the Dundee Resource.

Next steps include refactoring the `process_slivka` method and implementing an asynchronous workflow.