Upgrade jQuery and implement subresource integrity
Responding to another code alert this time regarding jQuery loading. The
recommendation was to use Subresource Integrity (SRI) to prevent
man-in-the-middle attacks. I also took the opportunity to upgrade the
jQuery version to the latest version for security and performance.

- Added Subresource Integrity (SRI) attribute to the jQuery script tag
- Upgraded jQuery version to 3.7.1. This is the latest version.

Addresses code scanning alert: Inclusion of functionality from an
untrusted source